rename: global 컨트롤러 패키지명 변경 #30
